% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{makeFlipbook}
\alias{makeFlipbook}
\title{Make a Flipbook in a shiny app}
\usage{
makeFlipbook(input, output, session, myplot)
}
\arguments{
\item{myplot}{a ggplot wrapped in `rlang::quo`}
}
\value{

}
\description{
Make a Flipbook in a shiny app
}
