% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{shinyflipbookr}
\alias{shinyflipbookr}
\title{Create a Shiny Flipbook}
\usage{
shinyflipbookr(appTitle = "shinyflipbookr example")
}
\arguments{
\item{appTitle}{The title to be displayed}
}
\value{

}
\description{
Create a Shiny Flipbook
}
